+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # CHANGELOG
 
+## v0.3.0 - 2021-05-13
+
+### fixed
+- Paths starting with `/` are assumed to be `earthengine-public` assets if no project is specified.
+- Use module logger level instead of root logger.
+
+### added
+- Add `tree` function to recursively list assets in folder
+- Add `export`, `exportImage`, and `exportTable` to export assets to cloud storage
+- Add `saveImage` and `findOrSaveImage` to export or cache images as assets
+
+### changed
+- Update `download` to work for feature collections
+- Update `download` to be able to recursively download image collections
+- Storage operations can take a bucket as an optional parameter instead of always using the default.
+- Storage operations will try to authenticate with default environment credentials the first time they are called.
+
+
 ## v0.2.3 - 2021-02-05
 
 ### fixed